Escaping Post

+++ admin/includes/view_all_users.php
@@ -91,11 +91,17 @@
         confirmQuery($change_to_sub_query);
         header("Location: users.php"); //Function to refresh the page
     }
-
+	
+	//preventing sql injection
     if(isset($_GET['delete'])){
-        $the_user_id = $_GET['delete'];
-        $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-        $delete_query = mysqli_query($connection, $query);
-        header("Location: users.php"); //Function to refresh the page
+		if(isset($_SESSION['user_role'])){
+			if($_SESSION['user_role'] == "admin"){
+				$the_user_id = mysql_escape_string($connection, $_GET['delete']);
+        		$query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+        		$delete_query = mysqli_query($connection, $query);
+        		header("Location: users.php"); //Function to refresh the page	
+			}
+		}
+       
     }
 ?>
